got rid of 'get' and 'set' special app functions, defined defactive
Now defapp still techically creates an implicity *state* variable
but you are free to ignore it. Instead of updating the DOM whenever
`set` is called, the auto-updates will happen after any call to
a function defined with `defactive`.

So you are free to put your state anywhere ,not just in *state*.

This was a hard stylistic decision to make: I chose to trust that
developers will not spread state too far and wide and to just use
defactive funcitons to be the units of dom-rerendering.